% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateToR.r
\name{aggregate_to_r}
\alias{aggregate_to_r}
\title{Aggregate to R}
\usage{
aggregate_to_r(x, dplyr = TRUE, nosave = FALSE, ...)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}

\item{nosave}{A value of FALSE processes the save commands (default),
a value of TRUE continues processing within R, overriding 
default x object. Extreme care with this feature.}

\item{...}{Additional arguments passed to function, not currently supported.}
}
\description{
Converts SPSS aggregate syntax to aggregate in R.
}
\details{
This function returns a matrix that highlights R syntax that mimics
the analysis done in SPSS.  The R syntax used by default is from the 
data.table package as this allows the ability to easily save aggregated 
variables back into the original R data frame.
}

