% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar.R
\name{bivar}
\alias{bivar}
\title{Bivariate table for multiple response questions}
\usage{
bivar(df, q, f, chi = T, excel = T)
}
\arguments{
\item{df}{Dataframe imported from SPSS file through `haven::read_sav()` function.}

\item{q}{Prefix for multiple response variable (ex: if there are 5 subquestions V1_1, v1_2, V1_3 etc. "V1" should be entered  into the function).}

\item{f}{Factor variable entered as a string ("V2")}
}
\description{
This function is aimed to make crosstab with multiple response question, coded in binary variables. Output table is saved in global environment under the name "table_factor_main_question".
}
