% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ComputeToR.r
\name{compute_to_r}
\alias{compute_to_r}
\title{Compute to R}
\usage{
compute_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default),
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS compute statements to variable creation in R.
}
\details{
This function returns a matrix that highlights R syntax that mimics
the analysis done in SPSS.
}

