% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correlations_to_r}
\alias{correlations_to_r}
\title{Correlations to R}
\usage{
correlations_to_r(x)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}
}
\description{
Converts SPSS correlations routine syntax to R correlation syntax
}
\details{
Generates a correlation matrix for the variables.
}

