% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrosstabToR.r
\name{crosstabs_to_r}
\alias{crosstabs_to_r}
\title{Crosstab to R}
\usage{
crosstabs_to_r(x, dplyr = TRUE, ...)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}

\item{...}{Additional arguments passed to function, not currently supported.}
}
\description{
Converts SPSS crosstab syntax to R table syntax.
}

