% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_to_r.r
\name{define_to_r}
\alias{define_to_r}
\title{Define to R}
\usage{
define_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS define statements to valid define statements in R.
}
\details{
This function returns a matrix that highlights R syntax that mimics
the analysis done in SPSS.
}

