% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorepeat_to_r.r
\name{dorepeat_to_r}
\alias{dorepeat_to_r}
\title{Do Repeat to R}
\usage{
dorepeat_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS do repeat statements to valid do repeat statements in R.
}
\details{
This function returns a matrix that highlights R syntax that mimics
the analysis done in SPSS.
}

