% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filehandle_to_r.r
\name{filehandle_to_r}
\alias{filehandle_to_r}
\title{File Handle to R}
\usage{
filehandle_to_r(x, dplyr = TRUE, ...)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}

\item{...}{Additional arguments passed to function, not currently supported.}
}
\description{
Converts SPSS file handle statements to valid file handle statements in R.
}
\details{
This function returns a matrix that highlights R syntax that mimics
the analysis done in SPSS.
}

