% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FrequenciesToR.r
\name{frequencies_to_r}
\alias{frequencies_to_r}
\title{Frequencies to R}
\usage{
frequencies_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default),
a value of FALSE uses data.table syntax}
}
\description{
Convert SPSS frequencies command to R syntax.
}

