% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToR.r
\name{get_to_r}
\alias{get_to_r}
\title{Get to R}
\usage{
get_to_r(x, dplyr = TRUE, nosave = FALSE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}

\item{nosave}{A value of FALSE processes the save commands (default),
a value of TRUE continues processing within R, overriding 
default x object. Extreme care with this feature as 
get commands will be ignored.}
}
\description{
Converts SPSS get syntax to R syntax
}

