% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getdata_to_r}
\alias{getdata_to_r}
\title{Get Data to R}
\usage{
getdata_to_r(x)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}
}
\description{
Converst SPSS Get Data command to R syntax.  Available for delimited or
excel data files.
}

