% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GraphToR.r
\name{graph_to_r}
\alias{graph_to_r}
\title{Graphics to R}
\usage{
graph_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default),
a value of FALSE uses data.table syntax}
}
\description{
Convert SPSS graph command to an R graph call using ggplot2.
}

