% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ComputeToR.r
\name{if_to_r}
\alias{if_to_r}
\title{If to R}
\usage{
if_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default),
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS if statements to R if statements.
}
\details{
This function returns R syntax that mimics what is accomplished
  with the same analysis in SPSS.
}

