% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchfiles_to_r.r
\name{matchfiles_to_r}
\alias{matchfiles_to_r}
\title{Match Files to R}
\usage{
matchfiles_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS match files statements to valid match files statements in R.
}
\details{
This function returns a matrix that highlights R syntax that mimics
the analysis done in SPSS.
}

