\name{minnesota}
\docType{data}
\alias{minnesota}
\title{Yearly College Football Data for University of Minnesota}
\format{A 1257 x 10 data frame}
\description{
  This data frame contains game by game college football
  from the year 1900 to 2012.
}
\section{Variables}{
 Variables: \itemize{ \item W.L, Win/Loss variable
 \item Date, Date of the game \item PF, Points For 
 \item PA, Points Against \item Opponent, 
 Opponent of the game \item Year, Year of the game
 \item Record, Yearly record \item conference, 
 Yearly conference affiliation \item PF.Year, 
 Points For each year \item PA.Year, 
 Points Against each year }
}
\keyword{datasets}
