% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/missingValueToR.r
\name{missingvalues_to_r}
\alias{missingvalues_to_r}
\title{Missing Values to R}
\usage{
missingvalues_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default),
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS missing value syntax to R code. Currently assumes
  that there is only one missing value code per line.
}

