% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaToR.r
\name{oneway_to_r}
\alias{oneway_to_r}
\title{Oneway to R}
\usage{
oneway_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS oneway syntax to R syntax
}

