% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print_rsyntax.R
\name{print.rsyntax}
\alias{print.rsyntax}
\title{Prints a rsyntax object}
\usage{
\method{print}{rsyntax}(x, ...)
}
\arguments{
\item{x}{The rsyntax object.}

\item{\ldots}{ignored}
}
\description{
Prints a rsyntax object.
}

