% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueLabelsToR.r
\name{renamevariables_to_r}
\alias{renamevariables_to_r}
\title{Rename Variables to R}
\usage{
renamevariables_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS rename variables syntax to R code.
}

