% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.r
\name{run_shiny}
\alias{run_shiny}
\title{Run Shiny Application Demo}
\usage{
run_shiny()
}
\description{
Function runs Shiny Application Demo
}
\details{
This function does not take any arguments and will run the Shiny Application.
If running from RStudio, will open the application in the viewer, otherwise will
use the default internet browser.
}

