% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SortCasesToR.r
\name{sortcases_to_r}
\alias{sortcases_to_r}
\title{Sort Cases to R}
\usage{
sortcases_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}
}
\description{
Convert SPSS sort cases command to an R function.
}

