% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTestToR.r
\name{ttest_to_r}
\alias{ttest_to_r}
\title{T-test to R}
\usage{
ttest_to_r(x, dplyr = TRUE, ...)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}

\item{...}{Additional arguments passed to function, not currently supported.}
}
\description{
Converts SPSS t-test syntax to R syntax
}

