% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggregate_to_r}
\alias{aggregate_to_r}
\title{Aggregate to R}
\usage{
aggregate_to_r(x)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}
}
\description{
Converts SPSS aggregate syntax to aggregate in R.
}
\details{
This function returns a matrix that highlights R syntax that mimics
the analysis done in SPSS.  The R syntax used by default is from the
data.table package as this allows the ability to easily save aggregated
variables back into the original R data frame.
}

