% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationsToR.r
\name{correlations_to_r}
\alias{correlations_to_r}
\title{Correlations to R}
\usage{
correlations_to_r(x, dplyr = TRUE, ...)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}

\item{...}{Additional arguments passed to function, not currently supported.}
}
\description{
Converts SPSS correlations routine syntax to R correlation syntax
}
\details{
Generates a correlation matrix for the variables.
}

