% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionToR.r
\name{regression_to_r}
\alias{regression_to_r}
\title{Regression to R}
\usage{
regression_to_r(x, dplyr = TRUE, ...)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default), 
a value of FALSE uses data.table syntax}

\item{...}{Additional arguments passed to function, not currently supported.}
}
\description{
Converts SPSS regression syntax to R syntax
}

