% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ValueLabelsToR.r
\name{valuelabels_to_r}
\alias{valuelabels_to_r}
\title{Value labels to R}
\usage{
valuelabels_to_r(x, dplyr = TRUE)
}
\arguments{
\item{x}{SPSS syntax - read in by SPSStoR function}

\item{dplyr}{A value of TRUE uses dplyr syntax (default),
a value of FALSE uses data.table syntax}
}
\description{
Converts SPSS value label syntax to R code#'
}

